+++ crates/ruma-events/tests/ui/07-enum-sanity-check.rs
@@ -6,10 +6,8 @@ event_enum! {
     events: [
         "m.direct",
         #[cfg(test)]
-        "m.dummy",
         "m.ignored_user_list",
         "m.push_rules",
-        "m.room_key",
         #[cfg(any())]
         "m.ruma_test",
     ]